Should solve glitches in time calculation see: https://confluence.desy.de/pages/resumedraft.action?draftId=137987362&draftShareId=3bfe073a-fb90-4ab7-8060-218a10b28547&

+++ pymepix/pymepix/processing/packetprocessor.py
@@ -138,11 +138,11 @@ class PacketProcessor(BasePipelineObject):
 
 
         if pixels.size > 0:
-            self.process_pixels(pixels, longtime)
+            self.process_pixels(np.int64(pixels), longtime)
 
         if triggers.size > 0:
             # print('triggers', triggers, longtime)
-            self.process_triggers(triggers, longtime)
+            self.process_triggers(np.int64(triggers), longtime)
 
         if self._handle_events:
